+++ src/api/consumer/lib/login.js
@@ -3,9 +3,9 @@ import loginToBitSrc from '../../../consumer/login/login';
 
 export default (async function loginAction(
   port: string,
-  noLaunchBrowser?: boolean,
+  suppressBrowserLaunch?: boolean,
   npmrcPath: string,
   skipRegistryConfig: boolean
 ): Promise<{ isAlreadyLoggedIn?: boolean, username?: string }> {
-  return loginToBitSrc(port, noLaunchBrowser, npmrcPath, skipRegistryConfig);
+  return loginToBitSrc(port, suppressBrowserLaunch, npmrcPath, skipRegistryConfig);
 });