+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - fix - operation aborted error when aborting username/pass authentication
 - fix handle tsx files when detectiveOption is empty
 - backward compatibility for components with environments with latest version
+- change no-launch-browser to suppress-browser-launch in bit login flag
 
 ## [13.0.0-dev.11] - 2018-06-13